+++ src/components/membership-components/AddSubscriptionDialog.tsx
@@ -17,7 +17,7 @@ import { Label } from '@/components/ui/label'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import { Textarea } from '@/components/ui/textarea'
 import { useBillingCycles } from '@/hooks/membership/useBillingCycles'
-import { useCreateSubscriptionPlan } from '@/hooks/membership/useCreateSubscriptionPlan'
+import { useCreateSubscriptionPlan } from '@/hooks/membership/useAddSubscriptionPlan'
 
 interface AddSubscriptionDialogProps {
   createdById: string | null
